combine image crawler into webshot
